content/blog: refactoring naming/numbering scheme
The idea here being that drafts are prefixed with `xxx_`, not knowing
when and in what order they will be released. This fixes the problem
where I start lots of articles, give them a number, then release them
in non-linear sequences and get confused down the line.

This commit _also_ releases the `nix-ocitools` article